+++ core-ng/src/main/java/core/framework/internal/kafka/MessagePublisherImpl.java
@@ -40,13 +40,12 @@ public class MessagePublisherImpl<T> implements MessagePublisher<T> {
     @Override
     public void publish(String topic, String key, T value) {
         if (topic == null) throw new Error("topic must not be null");
-        if (key == null) throw new Error("key must not be null");   // if key is null, kafka will pick random partition which breaks determinacy
 
         var watch = new StopWatch();
         validator.validate(value, false);
         byte[] message = mapper.toJSON(value);
         try {
-            var record = new ProducerRecord<>(topic, null, System.currentTimeMillis(), Strings.bytes(key), message, null);
+            var record = new ProducerRecord<>(topic, null, System.currentTimeMillis(), key == null ? null : Strings.bytes(key), message, null);
             linkContext(record.headers());
             producer.send(record);
         } finally {