+++ core-ng/src/main/java/core/framework/kafka/MessagePublisher.java
@@ -1,13 +1,12 @@
 package core.framework.kafka;
 
-import java.util.UUID;
-
 /**
  * @author neo
  */
 public interface MessagePublisher<T> {
+    // kafka uses sticky partitioning if key is null, refer to org.apache.kafka.clients.producer.internals.DefaultPartitioner
     default void publish(T value) {
-        publish(UUID.randomUUID().toString(), value);
+        publish(null, value);
     }
 
     void publish(String key, T value);