+++ core-ng/src/test/java/core/framework/internal/kafka/MessagePublisherImplTest.java
@@ -3,10 +3,12 @@ package core.framework.internal.kafka;
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
 import core.framework.kafka.MessagePublisher;
+import core.framework.util.Strings;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
+import java.util.Arrays;
 import java.util.List;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -38,13 +40,15 @@ class MessagePublisherImplTest {
         message.stringField = "value";
 
         publisher.publish(message);
-
         verify(producer).send(argThat(record -> {
-            assertThat(record.key()).hasSize(36);
+            assertThat(record.key()).isNull();
             assertThat(new String(record.headers().lastHeader(MessageHeaders.HEADER_CORRELATION_ID).value(), UTF_8)).isEqualTo("correlationId");
             assertThat(new String(record.headers().lastHeader(MessageHeaders.HEADER_REF_ID).value(), UTF_8)).isEqualTo(actionLog.id);
             return true;
         }));
+
+        publisher.publish("topic", "key", message);
+        verify(producer).send(argThat(record -> Arrays.equals(Strings.bytes("key"), record.key()) && "topic".equals(record.topic())));
 
         logManager.end("end");
     }